requirements.txt
requests
python-dotenv
mysql-connector-python
tabulate
